+++ CHANGELOG.md
@@ -4,7 +4,7 @@ This changelog track changes to the struqture project starting at version v1.0.0
 
 ## 1.3.1
 
-* Fixed bug allowing the construction of Hermitian operator products with annihilator index higher than creator index when there are leading equal indices.
+* Fixed bug allowing the construction of Hermitian operator products with annihilator index lower than creator index when there are leading equal indices.
 
 ## 1.3.0